+++ devslab-kit-access-api/src/main/java/kr/devslab/kit/access/PermissionChecker.java
@@ -1,0 +1,16 @@
+package kr.devslab.kit.access;
+
+public interface PermissionChecker {
+
+    boolean hasPermission(Permission permission);
+
+    boolean hasAnyPermission(Permission... permissions);
+
+    boolean hasAllPermissions(Permission... permissions);
+
+    default void check(Permission permission) {
+        if (!hasPermission(permission)) {
+            throw new PermissionDeniedException(permission);
+        }
+    }
+}