+++ devslab-kit-core/src/main/java/kr/devslab/kit/core/exception/DevslabKitException.java
@@ -1,0 +1,12 @@
+package kr.devslab.kit.core.exception;
+
+public class DevslabKitException extends RuntimeException {
+
+    public DevslabKitException(String message) {
+        super(message);
+    }
+
+    public DevslabKitException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}